fix location of analysiscript
